requirements.txt
# Core Deep Learning
tensorflow==2.10.0
tensorflow-lite==2.10.0
keras==2.10.0

# Data Processing
numpy==1.23.5
pandas==1.5.3
pillow==9.4.0
opencv-python==4.7.0.72
scikit-learn==1.2.1
scikit-image==0.19.3

# Data Augmentation
albumentations==1.3.0
imgaug==0.4.0

# Visualization
matplotlib==3.6.3
seaborn==0.12.2
plotly==5.13.0

# Model Analysis
tensorboard==2.10.0
keras-tuner==1.3.5

# Metrics & Evaluation
scikit-metrics==0.1.0

# Configuration
pyyaml==6.0
omegaconf==2.3.0

# Utilities
tqdm==4.64.1
click==8.1.3
python-dotenv==0.21.1

# Jupyter Notebooks
jupyter==1.0.0
ipykernel==6.21.1
ipywidgets==8.0.4

# Model Optimization
tensorflow-model-optimization==0.7.5

# Image Processing
imageio==2.25.0

# Logging & Monitoring
mlflow==2.1.1
wandb==0.13.9

# Testing
pytest==7.2.1
pytest-cov==4.0.0

# Code Quality
black==23.1.0
flake8==6.0.0
pylint==2.16.1

# Documentation
sphinx==5.3.0
sphinx-rtd-theme==1.2.0
